% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claude_api_call.R
\name{tag_claude}
\alias{tag_claude}
\title{Annotate Text Data Using Claude's Models}
\usage{
tag_claude(
  user_prompt,
  sys_prompt = "",
  model = "claude-v1",
  max_tokens_to_sample = 100,
  temperature = 1,
  api_key = base::Sys.getenv("CLAUDE_API_KEY"),
  api_url = "https://api.anthropic.com",
  verbose = TRUE,
  storage = TRUE,
  rate_limit = 0.6
)
}
\arguments{
\item{user_prompt}{A character vector containing the text data to be annotated.}

\item{sys_prompt}{A character string containing the instruction for the Claude model to follow when annotating the text data. Default is an empty string.}

\item{model}{The name of the Claude model to use for annotation. Default is "claude-v1".}

\item{max_tokens_to_sample}{The maximum number of tokens to generate in the response. Default is 100.}

\item{temperature}{Numeric. The temperature parameter for the Claude model, controlling the randomness of the generated annotations. Default is 1.}

\item{api_key}{The API key for authenticating with the Claude API. By default, this is read from the "CLAUDE_API_KEY" environment variable.}

\item{api_url}{The URL for the Claude API endpoint. Default is "https://api.anthropic.com".}

\item{verbose}{Logical. If TRUE, prints progress and error messages. Default is TRUE.}

\item{storage}{Logical. If TRUE, saves the annotated responses as a CSV file. Default is TRUE.}

\item{rate_limit}{Numeric. The rate limit for API requests in seconds. Default is 0.6s.}
}
\value{
A character vector containing the annotated responses from the Claude API.
The function also saves the annotated responses as a CSV file in the "LLMoutput/columnname/model.csv" path.
}
\description{
This function sends text data to Claude's API for annotation based on a provided system prompt.
The annotated responses are saved as a CSV file.
}
\examples{
\dontrun{
my_data <- c("Apple", "Tomato", "Broccoli")
annotated_data <- tag_claude(my_data, sys_prompt = "Which one is a fruit?")
}
}
