% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pplx_api_call.R
\name{tag_perplexity}
\alias{tag_perplexity}
\title{Annotate Text Data Using Perplexity's Models}
\usage{
tag_perplexity(
  user_prompt,
  sys_prompt = "Be precise and concise.",
  api_key = base::Sys.getenv("perplexity_key"),
  model = "mistral-7b-instruct",
  api_url = "https://api.perplexity.ai/chat/completions",
  verbose = TRUE
)
}
\arguments{
\item{user_prompt}{A character vector containing the text data to be annotated.}

\item{sys_prompt}{A character string containing the instruction for the Perplexity model to follow when annotating the text data. Default is "Be precise and concise.".}

\item{api_key}{The API key for authenticating with the Perplexity API. By default, this is read from the "perplexity_key" environment variable.}

\item{model}{The name of the Perplexity model to use for annotation. Default is "mistral-7b-instruct".}

\item{api_url}{The URL for the Perplexity API endpoint. Default is "https://api.perplexity.ai/chat/completions".}

\item{verbose}{Logical. If TRUE, prints progress and error messages. Default is TRUE.}
}
\value{
A character vector containing the annotated responses from the Perplexity API.
The function also saves the annotated responses as a CSV file in the "LLMoutput/columnname/model.csv" path.
}
\description{
This function sends text data to Perplexity's API for annotation based on a provided system prompt.
The annotated responses are saved as a CSV file.
}
\examples{
\dontrun{
my_data <- c("Apple", "Tomato", "Broccoli")
annotated_data <- tag_perplexity(my_data, sys_prompt = "Which one is a fruit?")
}
}
