% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aitag.R
\name{get_response}
\alias{get_response}
\title{Retrieve GPT responses for a dataframe}
\usage{
get_response(dataframe)
}
\arguments{
\item{dataframe}{The dataframe for which to retrieve GPT responses. The name of this dataframe
is used to locate the directory containing the saved RDS files.}
}
\value{
A tibble with columns:
\describe{
\item{id}{The "id" values from the original dataframe.}
\item{relevant_gpt}{The GPT-generated annotations, retrieved from the saved RDS files.}
}
}
\description{
This function retrieves the GPT responses saved as RDS files for a given dataframe. It
returns a tibble containing the "id" values from the original dataframe and the corresponding
GPT-generated annotations.
}
\examples{
\dontrun{
my_data <- data.frame(
id = 1:3,
text = c("Example text 1", "Example text 2", "Example text 3")
)

gpt_annotate(my_data, "Classify the sentiment of this text as positive, negative or neutral.")

annotations <- get_response(my_data)
}

}
