% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_response.R
\name{get_response}
\alias{get_response}
\title{Retrieve GPT responses for a column}
\usage{
get_response(column, model)
}
\arguments{
\item{column}{The column for which to retrieve GPT responses. The name of this column
is used to locate the directory containing the saved RDS files.}

\item{model}{The name of the GPT model used for annotation.}
}
\value{
A tibble with columns:
\describe{
\item{id}{The "id" values from the original column.}
\item{relevant_gpt}{The GPT-generated annotations, retrieved from the saved RDS files.}
}
}
\description{
This function retrieves the GPT responses saved as RDS files for a given column. It
returns a tibble containing the "id" values from the original column and the corresponding
GPT-generated annotations.
}
\examples{
\dontrun{
my_data <- data.frame(
id = 1:3,
text = c("The cake is tasty", "The cake is terrible", "The cake is okay")
)

gpt_annotate(my_data, "Classify the sentiment of this text as positive, negative or neutral.")

annotations <- get_response(my_data)
}

}
