% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai_api_call.R
\name{tag_gpt}
\alias{tag_gpt}
\title{Annotate a dataframe using OpenAI's GPT models}
\usage{
tag_gpt(
  column,
  instruction = "",
  temperature = 0,
  api_key = Sys.getenv("openai_key"),
  model = "gpt-3.5-turbo-0125",
  api_url = Sys.getenv("openai_url")
)
}
\arguments{
\item{column}{A character vector containing the text data to be annotated.}

\item{instruction}{A character string containing the instruction for the GPT model to follow
when annotating the text data. Default is an empty string.}

\item{temperature}{Numeric. The temperature parameter for the GPT model, controlling the
randomness of the generated annotations. Default is 0.}

\item{api_key}{The API key for authenticating with the OpenAI API. By default, this is read
from the "openai_key" environment variable.}

\item{model}{The name of the GPT model to use for annotation. Default is "gpt-3.5-turbo-0125".}

\item{api_url}{The URL for the OpenAI API endpoint. By default, this is read from the
"openai_url" environment variable.}
}
\value{
A data frame containing the annotated responses from the GPT API. The data frame has one column:
"aitag" (the annotated response from GPT).
The function also saves the annotated responses as a CSV file in the "LLMoutput/columnname/model/.csv" path.
}
\description{
This function sends the text data from a dataframe to OpenAI's GPT API for annotation,
based on a provided instruction. The GPT responses are saved as a CSV file.
}
\examples{
\dontrun{
my_data <- data.frame(
  id = c("1", "2", "3"),
  text = c("Example text 1", "Example text 2", "Example text 3")
)
annotated_data <- tag_gpt(my_data$id,
"Classify the sentiment of this text as positive, negative or neutral.")
}

}
